+++ src/main/scala/com/kaizen/api/Main.scala
@@ -2,12 +2,12 @@ package com.kaizen.api
 
 import caliban.Http4sAdapter
 import com.kaizen.api.graphql.Schema
-import com.kaizen.api.internal.actedIn.controller.ActedInController
-import com.kaizen.api.internal.actedIn.repository.ActedInRepository
-import com.kaizen.api.internal.actor.controller.ActorController
-import com.kaizen.api.internal.actor.repository.ActorRepository
-import com.kaizen.api.internal.movie.controller.MovieController
-import com.kaizen.api.internal.movie.repository.MovieRepository
+import com.kaizen.api.services.actedIn.controller.ActedInController
+import com.kaizen.api.services.actedIn.repository.ActedInRepository
+import com.kaizen.api.services.actor.controller.ActorController
+import com.kaizen.api.services.actor.repository.ActorRepository
+import com.kaizen.api.services.movie.controller.MovieController
+import com.kaizen.api.services.movie.repository.MovieRepository
 import zio.interop.catz._
 import zio.{App, Runtime, URIO, ZEnv, ZIO}
