+++ src/main/scala/com/kaizen/api/graphql/Actor.scala
@@ -1,7 +1,7 @@
 package com.kaizen.api.graphql
 
-import com.kaizen.api.internal.actedIn.controller._
-import com.kaizen.api.internal.actor._
+import com.kaizen.api.services.actedIn.controller._
+import com.kaizen.api.services.actor._
 import com.kaizen.api.graphql
 
 final case class Actor(
