--- src/main/scala/com/kaizen/api/internal/actedIn/ActedInData.scala
@@ -1,6 +1,0 @@
-package com.kaizen.api.internal.actedIn
-
-import com.kaizen.api.internal.actor.ActorId
-import com.kaizen.api.internal.movie.MovieId
-
-final case class ActedInData(actorId: ActorId, movieId: MovieId)
