+++ src/main/scala/com/kaizen/api/services/actedIn/ActedInData.scala
@@ -1,0 +1,6 @@
+package com.kaizen.api.services.actedIn
+
+import com.kaizen.api.services.actor.ActorId
+import com.kaizen.api.services.movie.MovieId
+
+final case class ActedInData(actorId: ActorId, movieId: MovieId)
