+++ src/main/scala/com/kaizen/api/services/movie/MovieData.scala
@@ -1,3 +1,3 @@
-package com.kaizen.api.internal.movie
+package com.kaizen.api.services.movie
 
 final case class MovieData(id: MovieId, title: MovieTitle)
