+++ src/main/scala/com/kaizen/api/services/movie/repository/InMemoryMovieRepository.scala
@@ -1,7 +1,7 @@
-package com.kaizen.api.internal.movie.repository
+package com.kaizen.api.services.movie.repository
 
-import com.kaizen.api.internal.RepositoryError
-import com.kaizen.api.internal.movie._
+import com.kaizen.api.services.RepositoryError
+import com.kaizen.api.services.movie._
 import zio.stm.TMap
 import zquery.DataSource
